+++ shacl_validation/src/store/sparql.rs
@@ -1,4 +1,5 @@
 use iri_s::IriS;
+use prefixmap::PrefixMap;
 use srdf::SRDFSparql;
 
 use crate::validate_error::ValidateError;
@@ -11,9 +12,8 @@ pub struct Endpoint {
 }
 
 impl Endpoint {
-    // TODO: Replace by from_path
-    pub fn new(path: &str) -> Result<Self, ValidateError> {
-        match SRDFSparql::new(&IriS::new_unchecked(path)) {
+    pub fn new(iri: &str, prefixmap: &PrefixMap) -> Result<Self, ValidateError> {
+        match SRDFSparql::new(&IriS::new_unchecked(iri), prefixmap) {
             Ok(store) => Ok(Self { store }),
             Err(_) => Err(ValidateError::SPARQLCreation),
         }